pyproject.toml
[tool.poetry]
name = "demo-python-drm"
version = "0.1.0"
description = ""
authors = ["natthasath <natthasath.sak@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "demo_python_drm"}]

[tool.poetry.dependencies]
python = "^3.10"
pycryptodome = "^3.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
